Fix import paths after merging renamed human-readable-name module
